use existing nodes itest
