--- java-client/src/test/java/org/opensearch/client/opensearch/integTest/httpclient5/StatsIT.java
@@ -1,5 +1,0 @@
-package org.opensearch.client.opensearch.integTest.httpclient5;
-
-import org.opensearch.client.opensearch.integTest.AbstractStatsIT;
-
-public class StatsIT extends AbstractStatsIT implements HttpClient5TransportSupport {}
